third_party/blink/web_tests/printing/page-break-margin-collapsed-expected.txt

Test for big margin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS: page number of "page1" is 1
PASS: page number of "page2" is 2
All tests passed

PASS successfullyParsed is true

TEST COMPLETE
